Track '_' binding in doctests more accurately
Currently '_' is handled by adding it to Checker.builtIns
while the doctest is being processed, causing it to be ignored
at every scope within the doctest.  However '_' is only valid
after an unassigned value in the outer most scope of the doctest.

'_' is only valid after an expression, either as an assignment
or unbound causing a return value to become a doctest example 'want'.

The new algorithm does not need to modify Checker.builtIns,
instead adding a '_' binding to the scope as needed.
It does not cover all cases, but leans in favour of adding a
'_' binding to prevent '_' being an undefined name.