Fix TypeError when processing relative imports
Fixes lp:1560134

aec68a784 added module names to error messages,
however it caused a TypeError for relative imports
that do not specify a module such as:

   from . import x

This fixes the TypeError, and also adds the necessary
leading dots for relative import error messages.

Add tests for various types of relative imports.